Used one service provider to load the others depending on Laravel's version
