+++ src/Support/Laravel/L4ServiceProvider.php
@@ -1,11 +1,12 @@
 <?php
 namespace Aloha\Twilio\Support\Laravel;
 
-use Aloha\Twilio\Manager;
-use Illuminate\Support\ServiceProvider as BaseServiceProvider;
+use Illuminate\Support\ServiceProvider as LaravelServiceProvider;
 
-class L4ServiceProvider extends BaseServiceProvider
+class L4ServiceProvider extends LaravelServiceProvider
 {
+    use ServiceProviderTrait;
+
     /**
      * Boot Method
      */
@@ -14,31 +15,5 @@ class L4ServiceProvider extends BaseServiceProvider
         // Register commands.
         $this->commands('twilio.sms', 'twilio.call');
         $this->package('aloha/twilio');
-    }
-
-    /**
-     * Register the service provider.
-     */
-    public function register()
-    {
-        // Register manager for usage with the Facade.
-        $this->app->singleton('twilio', function ($app) {
-            $config = $app['config']->get('services.twilio');
-            return new Manager($config['default'], $config['connections']);
-        });
-
-        // Define an alias.
-        $this->app->alias('twilio', 'Aloha\Twilio\Manager');
-
-        // Register Twilio Test SMS Command.
-        $this->app->singleton('twilio.sms', 'Aloha\Twilio\Commands\TwilioSmsCommand');
-
-        // Register Twilio Test Call Command.
-        $this->app->singleton('twilio.call', 'Aloha\Twilio\Commands\TwilioCallCommand');
-
-        // Register TwilioInterface concretion.
-        $this->app->singleton('Aloha\Twilio\TwilioInterface', function ($app) {
-            return $app->make('twilio')->defaultConnection();
-        });
     }
 }
