+++ src/Support/Laravel/L5ServiceProvider.php
@@ -1,0 +1,9 @@
+<?php
+namespace Aloha\Twilio\Support\Laravel;
+
+use Illuminate\Support\ServiceProvider as LaravelServiceProvider;
+
+class L5ServiceProvider extends LaravelServiceProvider
+{
+    use ServiceProviderTrait;
+}
